Extend Union resolution to match full names first in a new pass.
This makes things a bit safer, as two similar records differing
only in their namespace will now resolve correctly.

Add back name guards on Enum and Fixed resolution to ensure they
are compatible. While not critical for Fixed, it's important for
an Enum as completely different enumerations could match if there
is a default field specified.