Add validation function for Schemas.
It would be better in general to not need this and only allow
correct by construction Schemas, but it's a little tricky to keep
the API as tidy as it is if we do that.

At the moment, codec functions all just return a codec, but to
ensure that we didn't put a duplicate in the union, we would
need to return an either. This would contaminate the pipeline.

Likewise, it's convenient to just allow people to create names
inline when creating Codecs; if we forced them to validate them
at every step it would make things quite a lot more painful.

Fortunately, it's quite ok to write a Codec, and add a quick test
to your test suite that the Schema is indeed valid.