Improve how environments are managed by separating inlining and deconflicting.
An environment is now just simple lists of schemas which are available to be
substituted into reader and writer schemas.

This means that only recursive types use Named types when deconflicting, and
it's quite explicit what's a recursive value after the deconflict step.

Separating the environments also allows for incremental refactoring.